improve code index
